pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "biovision-ai"
version = "0.1.0"
description = "AI-powered dermatology decision-support system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pillow>=9.5.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.2.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "black>=23.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["biovision_ai*"]
